docker: just append args to the end of the command
Since the `docker build` context path is order independent we can just
add our arguments directly to the end.

This also fixes a problem where the build context was specified
somewhere other than the last argument and we are injecting
replacements (the replacements got injected in the wrong place, breaking
the build).