e2e: always compile dockersource bin
This makes sure we are always using the latest version.
If needed we could add some flag to use a different bin, but for now
this s fine.

Also adds some things to bust the test cache automatically if any source
changes.
The way this is handled is not ideal since if any file changes, even
non-source files changes it will bust the test cache, but this is a nice
shortcut to make `go test` just work on its own without having to
orchestrate it with make.